requirements.txt
# === Core API Server ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
pydantic==2.10.4

# === LLM / Agent ===
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14

# === Data Providers ===
yfinance==0.2.51
alpha_vantage==3.0.0
fredapi==0.5.2

# === Backtesting ===
backtrader==1.9.78.123

# === Vector Memory ===
faiss-cpu==1.9.0.post1
sentence-transformers==3.4.1

# === Data Processing ===
pandas==2.2.3
numpy==2.2.1
ta==0.11.0

# === Utilities ===
httpx==0.28.1
